Networked crafting and gathering
