+++ nyx/Cargo.toml
@@ -6,4 +6,5 @@ edition = "2021"
 [dependencies]
 bincode = "1.3.3"
 glam = { version = "0.26.0", features = ["bytemuck", "serde"] }
+rand = "0.8.5"
 serde = { version = "1.0.197", features = ["derive"] }
