+++ nyx/src/protocol.rs
@@ -1,5 +1,7 @@
 use glam::Vec3;
 
+use crate::item::ItemStack;
+
 pub const TPS: f32 = 20.0;
 
 #[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, serde::Deserialize, serde::Serialize)]
@@ -18,14 +20,17 @@ pub enum Clientbound {
     AuthSuccess(ClientId),
     Spawn(ClientId, Vec3),
     Despawn(ClientId),
-    Move(ClientId, Vec3, Tick)
+    Move(ClientId, Vec3, Tick),
+    SetStack(ItemStack)
 }
 
 #[derive(Clone, Copy, Debug, serde::Deserialize, serde::Serialize)]
 pub enum Serverbound {
     AuthRequest,
     Move(Vec3, Tick),
-    Disconnect
+    Disconnect,
+    Craft(usize),
+    Gather(usize)
 }
 
 #[derive(Clone, Debug, serde::Deserialize, serde::Serialize)]
